requirements.txt
python-telegram-bot==12.8
python-dotenv==1.0.0
pytz
APScheduler==3.6.3
requests==2.31.0
urllib3==1.26.18 